Remove a double-checked invariant
Now that isInvalid() is a semantic property, drop the assertion for this
invariant in the ASTVerifier.  This should also remove the last client
that wasn't registering the lazy resolver and expecting to pull any old
interface type out, so change a hack to an assertion to hopefully catch
future callers before we remove the LazyResolver entirely.