[pmo] Add an assert to make an invariant of pmo explicit.
Specifically, due to the way PMO promotes values when promoting values in ossa,
we know that:

1. The available value aggregator will always copy available values when
   promoting.

2. If any of those copy_values cover the entire type we are promoting (rather
   than being a subtype), then we know that all other copies must also have this
   property.

The reason why we know that this is true is that:

1. PMO does not handle enums (sum types). It only handles structs and
   tuples. This means that any types we promote can not be recursive.

2. If PMO finds an available value that covers the entire type, it will not
   handle along other pathes available values that initialize the memory
   location separately. This is as a result of PMO being intended to optimize
   temporaries introduced by SILGen that do not have this property.

Additionally one can see this by looking at the way that PMO aggregates values:
all tuples/struct aggregates are inserted at the load site. This means that the
algorithm would not be able to find a place to insert a phi to join together the
copy_value with split subtypes if the copy_value covered the entire loaded type.

The reason why I am asserting this invariant is I am going to use this to fix a
few issues in PMO around loops that can only occur when promoting in ossa.