[Constant Evaluator] Make symbolic closures, which are representations of closure literals in the constant evaluator, store the substitution map that was in the constant evaluator state when the closure creation site (partial_apply or thin_to_thick_function) that created the symbolic closure was evaluated.
A substitution map is a mapping from generic type parameters to types.
This map is updated and maintained by the constant evaluator as it
evaluates function calls. At a partial-application site, the substitution
map captures the mapping from the types of the captured arguments (which
could be generic types) to the types of the symbolic value they represent
(which has to be concrete types). The substitution map would be necessary
to fold the symbolic closure into SIL code that constructs the closure
it represents.