Move an access control check above an invalid check
Redeclaration checking has had a strong effect on the ordering of type checking.  This wouldn't be a problem, except the way we record the kind of dependency edges is strongly tied to the DeclContext at which a request to typecheck something came from.

In this case, if redeclaration checking checks the invalid bit, pattern bindings in other files will be typechecked which causes a cascading dependency edge to be added where previously a private edge existed.  Luckily, access checking doesn't depend on the interface type, so the move should be safe modulo access checking now seeing more invalid declarations.