Force the interface types of members before walking
Emulate the old validation order by resolving all the interface types of
the members of an aggregate or extension before walking the members
themselves.

This is an unfortunate necessity given that witness matching relied on
the previous validation order validating the aggregate's members before
descending into the bodies of accessors and functions.  Under the new
order we eagerly walk into the bodies of these declarations, so witness
matching sees partial results and either silently fails by picking the
wrong witness or loudly fails by complaining that requirements are
invalid or don't exist.

This is the cheap and easy thing to do for now.  If we can untangle the
cycle handling in witness matching, all of this can be removed.