[Visual C++] Provide alternative to array designators.
Even if they are supported in GCC and Clang, C99 array designators are
not part of the current C++ standard (they will be in C++20), so they
are not supported in Visual C++.

This commit provides an alternative for Visual C++ (which should work in
other compilers) which doesn't use array designators.

I think Clang 10 will start printing a warning about this incorrect
usage.