[windows] Avoid warning in header in test turning into fatal error.
In Windows, ClangImporter seems to execute with Microsoft compatibility,
and there are some more warnings that do not happen in Unix. One of them
seems to be -Wmicrosoft-enum-forward-reference which warns about forward
defined enums, like the one that happens in this header. Since the
-warnings-as-errors flag is now passed down to ClangImporter, this
warning will make this test fail.

Simply use the #pragma to disable the warning for this piece of code.