[pmo] Add an assert that when using SILSSAUpdater in ossa, we only have singular values if the underlying type is either trivial or we have a take.
The reason why this is true is that we will be inserting new copy_values for
each available value implying that we can never have such a singular value.

I also added two test cases that show that we have a singular value with the
trivial type and that works.