[windows] Quote argument for Windows compatibility.
In Windows the dollar doesn't introduce variables, so it doesn't need to
be escaped, and was creating an error when running the test on Windows.
Howerver, the unescaped dollar will be interpreted like a variable in
Unix shells, and will fail the test. The solution is quoting the inside
of the argument, which disables the variable substitution on Linux, and
doesn't seem to affect the Windows command line parser.