Sema: Skip override checking for members of constrained extensions
When a method in an extension of a class looks like an
override of a method from the base class, we emit a
diagnostic.

However due to a bug we used to skip this diagnostic for
certain members of constrained extensions.

Indeed it feels like we should not be doing this check
at all for members of constrained extensions, so lets
explicitly skip it, fixing a source compatibility problem
that was introduced when the unrelated bug was fixed.

Fixes <rdar://problem/57029805>, <https://bugs.swift.org/browse/SR-11740>.