build: require libdispatch on non-Darwin targets
libdispatch is already required for Foundation.  At this point
libdispatch is available on all the supported platforms.  We simply
enable SyntaxParserLib and SourceKit on all targets, allowing users to
disable them.  If enabled on non-Darwin platforms, the user must specify
`SWIFT_PATH_TO_LIBDISPATCH_SOURCE` to indicate where to find
libdispatch.  This is done implicitly by build-script, resulting in no
change for users.