Disable test until ownership lowering after diagnostics transition is complete.
I talked with Ravi (the maintainer) about this and he is cool. I am going to
provide a patch that he can review to the oslog specific pass that also
undisables the test.

This test was passing last week when I originally tested it so that suggests the
special optimization pass is pattern matching too specific of a pattern.