[test] Use -emit-sorted-sil and reorder CHECKs to match output.
In some cases (Android ARMv7) the SIL output wasn't following the output
of the source code, and the test failed. Use -emit-sorted-sil to use a
stable sorting for the SIL, and reorder the source to match the SIL
output.

The output is now the same order not depending on the value of
StripOwnershipAfterSerialization.